admins/sphinx-monitoring/requirements.txt
mysql==0.0.1
MySQL-python==1.2.5
pkg-resources==0.0.0
PyMySQL==0.8.1
PyYAML==3.12
